+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Major: Added `Account` class to handle account related functions
 - Major: Added template for classes
 - Major: Added `cron` to handle cronjobs
+- Minor: Moved `check-in` and `stamina` to `hoyo` folder
 
 ## 6/12/2023
 - Minor: Added `skipCheck` option to bypass `account.fired` check and will always send a notification.